backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==1.2.6
langgraph==1.0.7
langchain-google-genai==4.1.3
langchain-text-splitters
llama-index==0.10.34
llama-index-embeddings-gemini==0.1.7
llama-index-vector-stores-chroma==0.1.6
chromadb==0.4.24

pypdf==4.2.0
httpx==0.28.1
pydantic-settings==2.2.1
python-multipart==0.0.9
google-generativeai==0.7.2
python-dotenv==1.0.1
langchain-core==0.2.9
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1
PyPDF2==3.0.1
python-docx==1.1.2
aiofiles==23.2.1
pydantic==2.7.4
